cnn/proto_v1/requirements.txt
black==25.1.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
packaging==25.0
pathspec==0.12.1
pep8==1.7.1
pillow==11.3.0
platformdirs==4.3.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
ruff==0.12.2
six==1.17.0
sympy==1.14.0
tomli==2.2.1
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
triton==3.3.1
typing_extensions==4.14.1
